Create caesar-cipher.py
implementation of caesar-cipher cryptographic algorithm using python.

encryption:
it asks for data to be encrypted and key for encryption from user , later also asks for name of file where user wants to store encrypted data  and also name of file where user wants to store key. These files are of ".pickle" format which are not readable by humans so privacy is maintained.

Decryption:
while decryption we have to give input a .pickle file that contains encrypted data and also .pickle file that contains key for decryption.